+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Easier way import custom errors
 * Support for passing entire header into validate
 * Add forgot password method to request the reset of a users password
+* Add ability to update password by forgot password token
 
 ### Changed
 * Have InvalidRoleError extend ForbiddenError to return same information